pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neslter-stingray"
version = "1.1.0"
description = "NES-LTER Stingray data processing and visualization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.0",
    "matplotlib>=3.8",
    "scipy>=1.11",
    "plotly>=5",
    "dash>=3",
    "flask>=3",
    "werkzeug>=3",
    "jinja2>=3.1",
    "numba>=0.59",
    "PyYAML>=6",
    "tqdm>=4.64",
]

[project.scripts]
stingray = "stingray.cli.main:main"

[project.optional-dependencies]
dev = [
    "jupyter>=1",
    "ipykernel>=6.29",
    "pytest>=8",
]

shadowgraph = [
    "opencv-python",
    "tator",
]

all = [
    "jupyter>=1",
    "ipykernel>=6.29",
    "pytest>=8",
    "opencv-python",
    "tator",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]